# vail_scraper/routers/api/v2/openapi.yaml
openapi: "3.1.0"
info:
  title: "Vail scraper API"
  version: "2"
  summary: "Automatically scrapes vail's API and returns it in a usable format"
  description: |
    Rate limits are per-ip, however overrides may be added if you ask.
  license:
    name: "GPL v3"
    identifier: "GPL-3.0"
  contact:
    name: "Discord"
    url: "https://discord.com/users/397745647723216898"
servers:
  - description: "Production"
    url: "https://vail-scraper.farfrom.world/api/v2"
  - description: "Local instance"
    url: "http://localhost:8000"
...
paths:
...
  /users/{user_id}/stats:
    get:
      operationId: "getUserStats"
      summary: "Get user stats"
      description: |
          # Info
          Get someone's quick play stats
          
          # Notice
          This doesn't count customs.

          # Rate limits
          This endpoint has a limit of 5 requests per 5 seconds.
      tags: ["users"]
      
      parameters:
        - name: "user_id"
          in: "path"
          required: true
          schema:
            $ref: "#/components/schemas/UserId"
      
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
        "404":
          description: "User not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserNotFoundError"
        "429":
          $ref: "#/components/responses/RateLimited"
  /users/{user_id}/stats/timeseries:
    get:
      operationId: "getUserStatsTimeseries"
      summary: "Get user historical stats"
      description: |
          # Info
          Get someone's quick play stats from back in time
          
          # What stats are tracked
          This doesn't count customs.
          
          # Sorting of response items
          Providing `after`: lowest matching timestamp -> highest
        
          Providing `before`: highest matching timestamp -> lowest
        
          Providing nothing: most recent scrape -> oldest scrape
        

          # Rate limits
          This endpoint is currently not rate limited.
      tags: ["users"]
      
      parameters:
        - name: "user_id"
          in: "path"
          required: true
          schema:
            $ref: "#/components/schemas/UserId"
        - name: "before"
          in: "query"
          description: |
            This is mutually exclusive with `after`
            
            **NOTE** Please read the request description regarding sorting
          schema:
            $ref: "#/components/schemas/Timestamp"
            description: "if not sent this will default to the current timestamp"
        - name: "after"
          in: "query"
          description: |
            This is mutually exclusive with `before`
            
            **NOTE** Please read the request description regarding sorting
          schema:
            $ref: "#/components/schemas/Timestamp"
            default: 0
        - name: "limit"
          in: "query"
          schema:
            type: "number"
            default: 100
            minimum: 0
            maximum: 100
      
      responses:
        "200":
          description: "Success"
          content:
            application/json:
               schema:
                 type: "object"
                 required:
                   - "items"
                 properties:
                   items:
                     type: "array"
                     items:
                       $ref: "#/components/schemas/UserStatsTimeseriesEntry"
        "404":
          description: "User not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserNotFoundError"
...
components:
  responses:
    RateLimited:
      description: |
        You have been rate limited.
        
        If there is a legitimite reason you are consistently hitting this, please contact me and I will potentially grant an override.
      content:
        application/json:
          schema:
            type: "object"
            required:
              - "code"
              - "detail"  
            properties:
              code:
                type: "string"
                enum:
                  - "rate_limited"
              detail:
                type: "string"
                example: "rate limited"
  schemas:
...
    UserId:
      type: "string"
      format: "uuid-no-dashes"
      example: "086c0bdf4c8247eda847f876b49c057a"
...
    Timestamp:
      type: "number"
      format: "unix timestamp"
      example: 1717260094.5
    StringTimestamp:
      type: "string"
      format: "unix timestamp"
      example: "1717260094.5"
    
    # User stats
    UserStats:
      type: "object"
      required:
        - "maps"
        - "weapons"
        - "gamemodes"
        - "general"
        
      properties:
        maps:
          $ref: "#/components/schemas/UserMapsStats"
        weapons:
          $ref: "#/components/schemas/UserWeaponsStats"
        gamemodes:
          $ref: "#/components/schemas/UserGamemodeStats"
        general:
          $ref: "#/components/schemas/UserGeneralStats"
    UserStatsTimeseriesEntry:
      type: "object"
      required:
        - "maps"
        - "weapons"
        - "gamemodes"
        - "general"
        
      properties:
        maps:
          $ref: "#/components/schemas/UserMapsStats"
        weapons:
          $ref: "#/components/schemas/UserWeaponsStats"
        gamemodes:
          $ref: "#/components/schemas/UserGamemodeStats"
        general:
          $ref: "#/components/schemas/UserGeneralStats"
        timestamp:
          $ref: "#/components/schemas/StringTimestamp"
    UserMapsStats:
      type: "object"
      required:
        - "nine"
        - "thevoid"
        - "miru"
        - "suna"
        - "atmos"
        - "volt"
        - "cliffside"
        - "khidi"
        - "este"
        - "maar"
        
      properties:
        nine:
          $ref: "#/components/schemas/UserMapStats"
        thevoid:
          $ref: "#/components/schemas/UserMapStats"
        miru:
          $ref: "#/components/schemas/UserMapStats"
        suna:
          $ref: "#/components/schemas/UserMapStats"
        atmos:
          $ref: "#/components/schemas/UserMapStats"
        volt:
          $ref: "#/components/schemas/UserMapStats"
        cliffside:
          $ref: "#/components/schemas/UserMapStats"
        khidi:
          $ref: "#/components/schemas/UserMapStats"
        este:
          $ref: "#/components/schemas/UserMapStats"
        maar:
          $ref: "#/components/schemas/UserMapStats"
    UserMapStats:
      type: "object"
      required:
        - "match_results"
      properties:
        match_results:
          $ref: "#/components/schemas/UserStatsMapMatchResults"
    UserStatsMapMatchResults:
      type: "object"
      required:
        - "wins"
        - "losses"
        - "draws"
        - "abandons"
      properties:
        wins:
          type: "number"
        losses:
          type: "number"
        draws:
          type: "number"
        abandons:
          type: "number"
    UserWeaponsStats:
      type: "object"
      required:
        - "kanto"
        - "scarh"
        - "pl14"
        - "pm9"
        - "ump45"
        - "mk23"
        - "vityaz"
        - "g17"
        - "grotb"
        - "vector45"
        - "deagle"
        - "apc9pro"
        - "ak12"
        - "g28z"
        - "mr96"
        - "stf12"
        - "mk418"
        - "ak303n"
        - "m2rifled"
      properties:
        kanto:
          $ref: "#/components/schemas/UserStatsWeaponKanto"
        scarh:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        pl14:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        pm9:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        ump45:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        mk23:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        vityaz:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        g17:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        grotb:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        vector45:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        deagle:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        apc9pro:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        ak12:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        g28z:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        mr96:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        stf12:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        mk418:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        ak303n:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
        m2rifled:
          $ref: "#/components/schemas/UserStatsWeaponGunGeneric"
    UserStatsWeaponGunGeneric:
      type: "object"
      required:
        - "kills"
        - "shots"
      properties:
        kills:
          type: "object"
          required:
            - "total"
            - "headshot_kills"
          properties:
            total:
              type: "number"
            headshot_kills:
              type: "number"
        shots:
          type: "object"
          required:
            - "fired"
            - "hits"
          properties:
            fired:
              type: "number"
            hits:
              type: "object"
              required:
                - "head"
                - "body"
                - "arm"
                - "leg"
              properties:
                head:
                  type: "number"
                body:
                  type: "number"
                arm:
                  type: "number"
                leg:
                  type: "number"
    UserStatsWeaponKanto:
      type: "object"
      required:
        - "kills"
      properties:
        kills:
          type: "object"
          required:
            - "total"
            - "headshot_kills"
          properties:
            total:
              type: "number"
            headshot_kills:
              type: "number"
    UserGamemodeStats:
      type: "object"
      required:
        - "artifact"
        - "capture_the_orb"
        - "team_deathmatch"
        - "scoutzknivez"
        - "hardpoint"
        - "free_for_all"
        - "gun_game"
        - "one_in_the_chamber"
      properties:
        artifact:
          $ref: "#/components/schemas/UserGamemodeArtifactStats"
        capture_the_orb:
          $ref: "#/components/schemas/UserGamemodeCaptureTheOrbStats"
        team_deathmatch:
          $ref: "#/components/schemas/UserGamemodeTeamDeathmatchStats"
        scoutzknivez:
          $ref: "#/components/schemas/UserGamemodeScoutzknivezStats"
        hardpoint:
          $ref: "#/components/schemas/UserGamemodeHardpointStats"
        free_for_all:
          $ref: "#/components/schemas/UserGamemodeFreeForAllStats"
        gun_game:
          $ref: "#/components/schemas/UserGamemodeGunGameStats"
        one_in_the_chamber:
          $ref: "#/components/schemas/UserGamemodeOneInTheChamberStats"
...
    UserGeneralStats:
      type: "object"
      properties:
        time_played_seconds:
          type: "number"
        kills_and_deaths:
          type: "object"
          required:
            - "kills"
            - "assists"
            - "deaths"
            - "bursts"
          properties:
            kills:
              type: "number"
            assists:
              type: "number"
            deaths:
              type: "number"
            bursts:
              type: "object"
              required:
                - "2"
                - "3"
                - "5"
              properties:
                "2":
                  type: "number"
                "3":
                  type: "number"
                "5":
                  type: "number"
        match_results:
           type: "object"
           required:
             - "wins"
             - "draws"
             - "losses"
             - "abandons"
           properties:
             wins:
               type: "number"
             draws:
               type: "number"
             losses:
               type: "number"
             abandons:
               type: "number"
        prestige:
          type: "number"

              
    # User count
...
    UserNotFoundError:
      type: "object"
      required:
        - "code"
        - "detail"
      properties:
        code:
          type: "string"
          enum:
            - "user_not_found"
        detail:
          type: "string"
          example: "user not found"
...